refactor: move CsvFileGetter to business logic project
