+++ HerPortal.BusinessLogic/ExternalServices/CsvFiles/CsvFileGetter.cs
@@ -1,0 +1,75 @@
+﻿using HerPortal.BusinessLogic.ExternalServices.S3FileReader;
+using HerPortal.BusinessLogic.Models;
+using HerPortal.Data;
+using HerPublicWebsite.BusinessLogic.Services.S3ReferralFileKeyGenerator;
+
+namespace HerPortal.BusinessLogic.ExternalServices.CsvFiles;
+
+public class CsvFileGetter : ICsvFileGetter
+{
+    private readonly IDataAccessProvider dataAccessProvider;
+    private readonly S3ReferralFileKeyService keyService;
+    private readonly IS3FileReader s3FileReader;
+
+    public CsvFileGetter
+    (
+        IDataAccessProvider dataAccessProvider,
+        S3ReferralFileKeyService keyService,
+        IS3FileReader s3FileReader
+    ) {
+        this.dataAccessProvider = dataAccessProvider;
+        this.keyService = keyService;
+        this.s3FileReader = s3FileReader;
+    }
+    
+    public async Task<IEnumerable<CsvFileData>> GetByCustodianCodesAsync(IEnumerable<string> custodianCodes, int userId)
+    {
+        var downloads = await dataAccessProvider.GetLastCsvFileDownloadsAsync(userId);
+        var files = new List<CsvFileData>();
+
+        foreach (var custodianCode in custodianCodes)
+        {
+            var s3Objects = await s3FileReader.GetS3ObjectsByCustodianCodeAsync(custodianCode);
+            files.AddRange(s3Objects.Select(s3O =>
+                {
+                    var data = keyService.GetDataFromS3Key(s3O.Key);
+                    var downloadData = downloads.SingleOrDefault(d =>
+                        d.CustodianCode == data.CustodianCode
+                        && d.Year == data.Year
+                        && d.Month == data.Month
+                    );
+                    return new CsvFileData
+                    (
+                        data.CustodianCode,
+                        data.Month,
+                        data.Year,
+                        s3O.LastModified,
+                        downloadData?.LastDownloaded
+                    );
+                }
+            ));
+        }
+
+        return files
+            .OrderByDescending(f => new DateOnly(f.Year, f.Month, 1))
+            .ThenBy(f => LocalAuthorityData.LocalAuthorityNamesByCustodianCode[f.CustodianCode]);
+    }
+
+    public async Task<Stream> GetFileForDownloadAsync(string custodianCode, int year, int month, int userId)
+    {
+        if (!LocalAuthorityData.LocalAuthorityNamesByCustodianCode.ContainsKey(custodianCode))
+        {
+            throw new ArgumentOutOfRangeException(nameof(custodianCode), custodianCode,
+                "Given custodian code is not valid");
+        }
+
+        var fileStream = await s3FileReader.ReadFileAsync(custodianCode, year, month);
+        
+        // Notably, we can't confirm a download, so it's possible that we mark a file as downloaded
+        //   but the user has some sort of issue and doesn't get it
+        // We put this line as late as possible in the method for this reason
+        await dataAccessProvider.MarkCsvFileAsDownloadedAsync(custodianCode, year, month, userId);
+
+        return fileStream;
+    }
+}
