+++ HerPortal/Controllers/HomeController.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using HerPortal.BusinessLogic.ExternalServices.CsvFiles;
 using HerPortal.DataStores;
-using HerPortal.ExternalServices.CsvFiles;
 using HerPortal.Helpers;
 using HerPortal.Models;
 using Microsoft.AspNetCore.Mvc;
